requirements.txt
python-pptx
python-docx
customtkinter
windnd
